refactor: #964: Move useful subject import-related classes to new Common project.
