+++ Atlas.ManualTesting.Common/SubjectImport/ImportedSubject.cs
@@ -1,0 +1,19 @@
+﻿// ReSharper disable InconsistentNaming
+
+namespace Atlas.ManualTesting.Common.SubjectImport
+{
+    public class ImportedSubject
+    {
+        public string ID { get; set; }
+        public string A_1 { get; set; }
+        public string A_2 { get; set; }
+        public string B_1 { get; set; }
+        public string B_2 { get; set; }
+        public string C_1 { get; set; }
+        public string C_2 { get; set; }
+        public string DQB1_1 { get; set; }
+        public string DQB1_2 { get; set; }
+        public string DRB1_1 { get; set; }
+        public string DRB1_2 { get; set; }
+    }
+}
