+++ Atlas.MatchPrediction.Test.Validation/Atlas.MatchPrediction.Test.Validation.csproj
@@ -10,7 +10,6 @@
 
   <ItemGroup>
     <PackageReference Include="AzureExtensions.Swashbuckle" Version="3.2.2" />
-    <PackageReference Include="CsvHelper" Version="12.1.2" />
     <PackageReference Include="Microsoft.Azure.Functions.Extensions" Version="1.1.0" />
     <PackageReference Include="Microsoft.Azure.WebJobs.Extensions.EventGrid" Version="3.1.0" />
     <PackageReference Include="Microsoft.Azure.WebJobs.Extensions.ServiceBus" Version="5.3.0" />
@@ -19,6 +18,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Atlas.ManualTesting.Common\Atlas.ManualTesting.Common.csproj" />
     <ProjectReference Include="..\Atlas.MatchPrediction.Test.Validation.Data\Atlas.MatchPrediction.Test.Validation.Data.csproj" />
     <ProjectReference Include="..\Atlas.MatchPrediction\Atlas.MatchPrediction.csproj" />
   </ItemGroup>
