+++ Atlas.MatchPrediction.Test.Validation/Models/ImportedSubjectExtensions.cs
@@ -1,25 +1,11 @@
 ﻿
 // ReSharper disable InconsistentNaming
 
+using Atlas.ManualTesting.Common.SubjectImport;
 using Atlas.MatchPrediction.Test.Validation.Data.Models;
 
 namespace Atlas.MatchPrediction.Test.Validation.Models
 {
-    internal class ImportedSubject
-    {
-        public string ID { get; set; }
-        public string A_1 { get; set; }
-        public string A_2 { get; set; }
-        public string B_1 { get; set; }
-        public string B_2 { get; set; }
-        public string C_1 { get; set; }
-        public string C_2 { get; set; }
-        public string DQB1_1 { get; set; }
-        public string DQB1_2 { get; set; }
-        public string DRB1_1 { get; set; }
-        public string DRB1_2 { get; set; }
-    }
-
     internal static class ImportSubjectExtensions
     {
         public static SubjectInfo ToSubjectInfo(this ImportedSubject importedSubject, SubjectType subjectType)
